chore(developer): Further AddWarning and SetError cleanup
* Splits compiler interface functions that return bool into
  CompilerInterfaces.cpp, and moves `SetError` define to that unit so
  it can't be used accidentally elsewhere
* Audit of `AddWarning` calls and only those that are used in bool
  functions now `return FALSE` (with new `AddWarningBool`) whereas all
  others `return CERR_Break`
