+++ developer/src/kmcmplib/src/meson.build
@@ -37,6 +37,7 @@ lib = library('kmcmplib',
   'CasedKeys.cpp',
   'CharToKeyConversion.cpp',
   'Compiler.cpp',
+  'CompilerInterfaces.cpp',
   'DeprecationChecks.cpp',
   'Edition.cpp',
   'filesystem.cpp',
